Rename folder to SortFilterSearch
